backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
httpx==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0
supabase==2.3.0
anthropic==0.20.0
groq==0.9.0
openai==1.12.0
google-generativeai==0.3.2
python-dotenv==1.0.0
beautifulsoup4==4.12.0
lxml==5.1.0
rapidfuzz==3.6.0
pandas==2.2.0
tenacity==8.2.0
python-dateutil==2.9.0
slowapi==0.1.9
psycopg2-binary==2.9.9
playwright==1.40.0